Write code for section 4.6.2.
Add asynchronous tests. This test is actually a "mock"; that is, I
create a function that simulates an asynchronous call to test if a year
is a leap year. The test code invokes this function in an `async` macro.
This macro runs its tests and invokes the `done` method (passed in by
the test framework) when all the asynchronous tests are finished.